moved only_one_or_zero_is_set to app/models.py renamed test file for order_item_constraints
