+++ src/orders/models/order.py
@@ -1,39 +1,17 @@
 from typing import Iterable, Optional
 
-import operator
-from django.db.models import CheckConstraint, Q, QuerySet
+from django.db.models import CheckConstraint, QuerySet
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from django.utils.translation import pgettext_lazy
-from functools import reduce
 
-from app.models import TimestampedModel, models
+from app.models import TimestampedModel, models, only_one_or_zero_is_set
 from orders.fields import ItemField
 from products.models import Product
 
 
 class UnknownItemException(Exception):
     pass
-
-
-def only_one_or_zero_is_set(*fields: str) -> Q:
-    constraints = []
-    for field in fields:
-        constraint = Q(
-            **{
-                f'{field}__isnull': False,
-                **{f'{empty_field}__isnull': True for empty_field in fields if empty_field != field},
-            },
-        )
-        constraints.append(constraint)
-
-    all_fields_can_empty_constraint = Q(
-        **{f'{field}__isnull': True for field in fields},
-    )
-
-    constraints.append(all_fields_can_empty_constraint)
-
-    return Q(reduce(operator.or_, constraints))
 
 
 class OrderQuerySet(QuerySet):
